pyproject.toml
[tool.poetry]
name = "npp-load-factor-optimization"
version = "0.1.0"
description = ""
authors = ["alex <melnikov1991@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
ruff = "^0.12.1"
oemof-solph = "^0.6.0"
numpy = "^2.3.1"
pandas = "^2.3.0"
loguru = "^0.7.3"
oemof-visio = "^0.0.2a2"
tsam = "^2.3.9"
tabulate = "^0.9.0"
pathlib = "^1.0.1"
graphviz = "^0.21"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
